refactor(core): split order item input conversion
- move cart/resolved-item to OrderItemInput conversion into a focused helper module
- keep orderItems centered on menu lookup, item resolution, and quote aggregation
- reduce orderItems.ts hotspot score from 24.1 to 16.7 and density from 0.13 to 0.09
